22,March 2025-Initial Commit

+++ src/main/java/com/example/firstjobapp/job/impl/JobServiceImpl.java
@@ -15,7 +15,7 @@ public class JobServiceImpl implements JobService {
 
     //private List<Job> jobs=new ArrayList<>();
     JobRepository jobRepository;
-    private Long nextId=1L;
+    
 
     public JobServiceImpl(JobRepository jobRepository) {
         this.jobRepository = jobRepository;
@@ -28,7 +28,6 @@ public class JobServiceImpl implements JobService {
 
     @Override
     public void createJob(Job job) {
-     job.setId(nextId++);
      jobRepository.save(job);
     }
 
@@ -39,13 +38,13 @@ public class JobServiceImpl implements JobService {
 
     @Override
     public boolean deleteJobById(Long id) {
-        try {
+        if (jobRepository.existsById(id)) {  // Check if job exists
             jobRepository.deleteById(id);
             return true;
-        }catch (Exception e){
-            return false;
         }
+        return false; // Return false if job does not exist
     }
+
 
     @Override
     public boolean updateJob(Long id, Job updatedJob) {
@@ -58,6 +57,8 @@ public class JobServiceImpl implements JobService {
                 job.setMinSalary(updatedJob.getMinSalary());
                 job.setMaxSalary(updatedJob.getMaxSalary());
                 job.setLocation(updatedJob.getLocation());
+
+                jobRepository.save(job);
                 return true;
             }
         return false;
